adding forgotten files and finishing the TCP protocol in cpp
in this changes I have add the forgotten .h files, and finished the DLL with the establishment of  "tcp_communication.h" .h file and the "tcp_communication.cpp" .cpp file. furthermore, I'm not sure about the protocol usage of such script as it dose not have the akg and handshake that the TCP protocol reraise on.